pyproject.toml
[project]
name = "dr-holmes"
version = "0.1.0"
description = "Multi-agent differential diagnosis system"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.30.0",
    "langgraph>=0.2.0",
    "langchain-core>=0.2.0",
    "chromadb>=0.5.0",
    "datasets>=2.20.0",
    "pydantic>=2.7.0",
    "rich>=13.7.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
    "sentence-transformers>=3.0.0",
    "tiktoken>=0.7.0",
    "sqlalchemy>=2.0.0",
    "neo4j>=5.0.0",
    "redis>=5.0.0",
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.29.0",
    "websockets>=12.0",
    "asyncpg>=0.29.0",
    "aiosqlite>=0.19.0",
    "prometheus-client>=0.20.0",
    "python-jose[cryptography]>=3.3.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["dr_holmes"]
